[FLINK-7811] Add explicit types in Scala code
This is in preparation for adding support for Scala 2.12. We need to add
explicit types because the Scala 2.12 type checker will not work in
certain situations. Because pre-2.12 Scala did not have compatibility
between Scala lambdas and Java 8 Lambdas we have this construct in parts
of the API where we want to allow passing in Scala lambdas:

def flatMap(flatMapper: FlatMapFunction[T, R]): DataSet[R]
def flatMap(fun: (T, Collector[R]) => Unit): DataSet[R]

The "flatMap" method in FlatMapFunction has the same signature as the
lambda. With this Scala 2.12 can not decide which method to pick and the
type checker will not check the types of a lambda. Explicitly specifying
the types seems to work, though.